+++ .ci/default.nix
@@ -1,34 +1,26 @@
 { }:
 let
-  nixpkgsSets = import ./nixpkgs.nix;
-  inherit (nixpkgsSets) nixos1809 nixos2003 unstable;
-  inherit (unstable) lib;
-  inherit (nixos2003.haskell.lib) doJailbreak dontCheck markUnbroken overrideCabal;
-  ghcs = rec {
-    ghc865 = nixos2003;
-    ghc884 = nixos2003;
-    ghc8107 = unstable;
-    ghc902 = unstable;
+  nixos2311 = import ./nixpkgs/23.11 {};
+  pkgs = nixos2311;
+  inherit (pkgs) lib;
+  inherit (pkgs.haskell.lib) doJailbreak dontCheck markUnbroken overrideCabal;
+  ghcs = {
+
+    "ghc8107" = pkgs;
+    "ghc902" = pkgs;
+    "ghc928" = pkgs;
+    "ghc948" = pkgs;
+    "ghc963" = pkgs;
+    "ghc981" = import ./nixpkgs/unstable {};
   };
-  build = 
-    { nixpkgs ? import ./dep/nixpkgs {}
+  build =
+    { nixpkgs
     , ghc ? null
     }:
     let
       baseHaskellPackages = if ghc != null then nixpkgs.haskell.packages.${ghc} else nixpkgs.haskellPackages;
       haskellPackages = baseHaskellPackages.override {
-        overrides = self: super: import ../. { haskellPackages = self; } // {
-          which = self.callHackageDirect {
-            pkg = "which";
-            ver = "0.2.0.1";
-            sha256 = "0kfbzaz1wgzmyqvw4m57yj43x4ihalk5a7y703fa0pjv0cvvx0ss";
-          } {};
-          shelly = self.callHackageDirect {
-            pkg = "shelly";
-            ver = "1.9.0";
-            sha256 = "1x9d86pswkncyhnzpbx4a1kmn847kjqs0ivishn84h0w6lpf12pc";
-          } {};
-        };
+        overrides = self: super: import ../. { haskellPackages = self; };
       };
     in {
       inherit (haskellPackages)