+++ gargoyle-postgresql-connect/ChangeLog.md
@@ -1,5 +1,9 @@
 # Revision history for gargoyle-postgresql-connect
 
+## 0.1.0.2
+
+* Require resource-pool 0.4.*
+
 ## 0.1.0.1
 
 * Add support for ghc 9.0.2