+++ gargoyle/ChangeLog.md
@@ -1,5 +1,9 @@
 # Changelog for gargoyle
 
+## 0.1.2.0
+
+* Fix a bug with shutting down on Apple Silicon
+
 ## 0.1.1.1
 
 * Add support for ghc 9.0.2